refactor: Extract hasher package
