+++ forst/go.mod
@@ -1,9 +1,10 @@
 module forst
 
-go 1.24.0
+go 1.24.3
 
 require (
-	github.com/fsnotify/fsnotify v1.8.0 // indirect
-	github.com/sirupsen/logrus v1.9.3 // indirect
-	golang.org/x/sys v0.13.0 // indirect
+	github.com/fsnotify/fsnotify v1.8.0
+	github.com/sirupsen/logrus v1.9.3
 )
+
+require golang.org/x/sys v0.13.0 // indirect
